+++ api/controllers/api/admin/PostController.js
@@ -79,5 +79,21 @@ module.exports = {
       sails.log.error(e);
       res.serverError(e);
     }
-  }
+  },
+
+  changePublish: async (req, res) => {
+    try {
+      const { id } = req.params;
+      const { publish } = req.body;
+      res.ok({
+        message: 'update post success.',
+        data: await Post.update( { publish } , {
+          where: { id }
+        }),
+      });
+    } catch (e) {
+      sails.log.error(e);
+      res.serverError(e);
+    }
+  },
 }